requirements.txt
streamlit==1.50.0
pandas==2.3.3
numpy==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
matplotlib==3.10.7
seaborn==0.13.2
python-dateutil==2.9.0.post0
